+++ app/src/main/java/com/winterparadox/themovieapp/common/beans/Favorite.java
@@ -14,7 +14,7 @@ public class Favorite {
     public long time;
 
     @PrimaryKey
-    public int movieId;
+    public long movieId;
 
     public Favorite () {
     }
