+++ app/src/main/java/com/winterparadox/themovieapp/movieDetails/MovieDetailsApiService.java
@@ -12,6 +12,6 @@ import retrofit2.http.Path;
 public interface MovieDetailsApiService {
 
     @GET("movie/{id}?&append_to_response=release_dates,credits,similar")
-    Single<Movie> movieDetails (@Path("id") int id);
+    Single<Movie> movieDetails (@Path("id") long id);
 
 }
